Crate lock_keeper

Cryptography, protocols, and other shared types and context used by multiple entities in the Lock Keeper digital asset management system.

⚠️ Usage: This crate should not be used directly by developers outside of Bolt Labs. It is designed as a dependency for other crates in the ecosystem, including lock-keeper-key-server and lock-keeper-client.

TESTING

Re-exports

pub use error::LockKeeperError;

Modules

audit_event

Audit events, and associated fields and types

blockchain

Blockchain abstraction.

channel

config

crypto

Application-specific cryptographic types and operations.

defaults

error

keys

Digital asset keys and descriptions.

opaque_storage

pem_utils

rpc

timeout

transaction

Library for blockchains, transactions, and requests for signatures on transactions.

types

user

Models for the first pass of MongoDB integration.

Macros

impl_message_conversion

Generates TryFrom implementations to and from the Message type for a given list of types.

Enums

ClientAction

Options for actions the Lock Keeper client can take.

RetrieveContext

Options for the asset owner’s intended use of a secret

TestLogs

Logs used to verify that an operation completed in the integration tests.