Type Definition lock_keeper::channel::ServerChannel

pub type ServerChannel = Channel<Result<Message, Status>>;

Implementations

impl ServerChannel

pub fn create(
receiver: Streaming<Message>
) -> (Self, Receiver<Result<Message, Status>>)

Create a server-side Channel that sends error codes in addition to Message objects.

pub async fn send(
&mut self,
message: impl TryInto<Message>
) -> Result<(), LockKeeperError>

Send a message across the channel. This function accepts any type that can be converted to a `Message.

pub async fn send_error(
&mut self,
status: impl Into<Status>
) -> Result<(), LockKeeperError>