Module lock_keeper::crypto

Application-specific cryptographic types and operations.

Defines and implements keys and secret types, and the appropriate transformations between them. Public functions here are mostly wrappers around multiple low-level cryptographic steps.

Modules

arbitrary_secret 🔒

generic 🔒

signing_key 🔒

storage_key 🔒

Structs

Encrypted

A ciphertext representing an object of type T, encrypted under the ChaCha20Poly1305 scheme for authenticated encryption with associated data (AEAD).

Export

Raw material for an exported signing key.

Import

Raw material for an imported signing key.

KeyId

Universally unique identifier for a stored secret or signing key. Wrapped in a Box to avoid stack overflows during heavy traffic. KeyIds are created by implementors of our DataStore trait. So we expose the internal as pub.

MasterKey

The master key is a default-length symmetric encryption key for an AEAD scheme.

OpaqueSessionKey

A session key is produced as shared output for client and server from OPAQUE.

RemoteStorageKey

The remote storage key is a default-length symmetric encryption key for an AEAD scheme.

Secret

An arbitrary secret.

SignableBytes

Wrapper used to declare arbitrary bytes as Signable.

Signature

A signature on an object encrypted under the ECDSA signature scheme.

SigningKeyPair

An ECDSA signing key pair, including a public component for verifying signatures, a private component for creating them, and context about the key pair.

SigningPublicKey

The public component of an ECDSA signing key, and context about the key.

StorageKey

A storage key is a default-length symmetric encryption key for an AEAD scheme. The storage key is used to encrypt stored secrets and signing keys.

Enums

CryptoError

Errors that arise in the cryptography module.

Constants

_DERIVE_Drop_FOR_Export 🔒

_DERIVE_Drop_FOR_MasterKey 🔒

_DERIVE_Drop_FOR_OpaqueSessionKey 🔒

Traits

Signable

Provides the methods necessary to sign and verify a piece of data with a SigningKeyPair. This trait should be explicitly implemented on types that are intended to be signed.