Struct lock_keeper::crypto::cryptor::CryptorContext

pub struct CryptorContext {
    key_server_name: Vec<u8>,
}

The context (a.k.a. associated data).

TODO: We should consider making this a trait and require the calling application to define appropriate context for each of their encrypted types. See issue key-mgmt#542

Fields
key_server_name: Vec<u8>
Implementations

impl CryptorContext

pub fn new<T: AsRef<[u8]>>(key_server_name: T) -> Self

Constructs an instance of CryptorContext. A type parameter is bounded by the AsRef<[u8]> trait.

This trait is implemented by types that can be referenced as a byte slice, e.g. &str, Vec<u8>

Example usage:

From string slices…

    let context = CryptorContext::new("my_key_server");

From Vec<u8>…

    let key_server_name = "my_key_server".as_bytes().to_vec();
    let context = CryptorContext::new(key_server_name);