Enum lock_keeper::crypto::generic::CryptoError

pub enum CryptoError {

    CannotEncodeDataLength,
    CombineShardsFailed(String),
    ConversionError,
    DecryptionFailed,
    EncryptionFailed,
    InvalidEncryptionKey,
    SensitiveInfoCheckFailed,
    Infallible(Infallible),
    KeyDerivationFailed(InvalidLength),
    NonZeroScalarConversion,
    RandomNumberGeneratorFailed,
    IncorrectSealKeySize(usize),
    Signature(Error),
    ShardDecryptionFailed(String),
    ShardEncryptionFailed(String),
    ShardingFailed(String),
    VerificationFailed,
    FileIo(Error, PathBuf),

}

Errors that arise in the cryptography module.

Variants

CannotEncodeDataLength

Length of data too large for our integer data type.

CombineShardsFailed(String)

ConversionError

DecryptionFailed

EncryptionFailed

InvalidEncryptionKey

SensitiveInfoCheckFailed

Infallible(Infallible)

The impl<T> Encrypted<T> has some trait bounds for converting a TryFrom::Error associated type into a CryptoError. Rust automatically implements TryFrom<T, Error=Infallible> when From<T> is implemented. This variant ensures impl From<CryptoError> for Infallible holds true in this case.

KeyDerivationFailed(InvalidLength)

NonZeroScalarConversion

RandomNumberGeneratorFailed

IncorrectSealKeySize(usize)

Signature(Error)

ShardDecryptionFailed(String)

ShardEncryptionFailed(String)

ShardingFailed(String)

VerificationFailed

FileIo(Error, PathBuf)

IO error specific to file IO failing. Allows us to include the file that failed as part of the error.