Module lock_keeper::crypto::signing_key

Modules

generation_types

Structs

Import

Raw material for an imported signing key.

SignableBytes

Wrapper used to declare arbitrary bytes as Signable.

Signature

A signature on an object encrypted under the ECDSA signature scheme.

SigningKeyPair

An ECDSA signing key pair, including a public component for verifying signatures, a private component for creating them, and context about the key pair.

SigningPublicKey

The public component of an ECDSA signing key, and context about the key.

Constants

_DERIVE_Drop_FOR_Import 🔒

_DERIVE_Drop_FOR_SigningKeyPair 🔒

Traits

Signable

Provides the methods necessary to sign and verify a piece of data with a SigningKeyPair. This trait should be explicitly implemented on types that are intended to be signed.