Trait lock_keeper::crypto::signing_key::Signable

pub trait Signable: AsRef<[u8]> {
    fn sign(&self, signing_key: &SigningKeyPair) -> Signature;
    fn verify(
        &self,
        public_key: &SigningPublicKey,
        signature: &Signature
    ) -> Result<(), LockKeeperError>;
}

Provides the methods necessary to sign and verify a piece of data with a SigningKeyPair. This trait should be explicitly implemented on types that are intended to be signed.

It requires AsRef<[u8]> so that we can accept impl Signable in the public client API but send SignableBytes over the network. This requirement may be removed in the future.

Required Methods

fn sign(&self, signing_key: &SigningKeyPair) -> Signature

fn verify(
&self,
public_key: &SigningPublicKey,
signature: &Signature
) -> Result<(), LockKeeperError>