Struct lock_keeper::crypto::CryptorKey

pub struct CryptorKey {
    pub(super) key_material: Box<Key>,
    config: SensitiveInfoConfig,
}

The CryptorKey type is a default-length symmetric encryption key for an AEAD scheme. It can be used to securely encrypt data.

Fields
key_material: Box<Key>config: SensitiveInfoConfig
Implementations

impl CryptorKey

Implementation for CryptorKey

const CRYPTOR_KEY_LENGTH: usize = 32usize

Length of the encryption key in bytes

pub fn new(rng: &mut impl CryptoRng + RngCore) -> Self

Constructs a new instance of the CryptorKey type.

Arguments

rng - A mutable reference to an object implementing both the CryptoRng and RngCore traits. This is used to generate the cryptographic key for the ChaCha20Poly1305 algorithm.

Returns

Returns a new instance of the CryptorKey type with a fresh cryptographic key generated via the provided random number generator (rng) and a new SensitiveInfoConfig object.

A CryptorKey is generated uniformly at random. It is a 32-byte pseudorandom key for use in the ChaCha20Poly1305 scheme for authenticated encryption with associated data (AEAD).

pub fn from_file(path: impl AsRef<Path>) -> Result<Self, CryptoError>

Returns the CryptorKey stored in a file.

pub fn from_bytes(cryptor_key: &[u8]) -> Result<Self, CryptoError>

Returns the CryptorKey key based on key material in the caller’s byte array.

pub fn into_bytes(self) -> [u8; 32]

Converts CryptorKey to a byte array.

Warning

This method gives direct access to the key material bytes. The caller should be careful to manually zeroize them after use to prevent unintended exposure of sensitive information. Consider using the zeroize crate to securely zero the data.

Example

use zeroize::Zeroize;
use lock_keeper::crypto::CryptorKey;///

let mut rng = rand::thread_rng();
let encryption_key = CryptorKey::new(&mut rng);
let mut key_bytes = encryption_key.into_bytes();

// Use the key bytes...

// When done, zeroize the key bytes
key_bytes.zeroize();