Module lock_keeper::transaction

Library for blockchains, transactions, and requests for signatures on transactions.

Structs

AssetId

Unique ID for a digital asset.

TarId

Unique ID for a TransactionApprovalRequest.

Transaction

A transaction describes the transfer of a digital asset owned by an asset owner to another entity.

TransactionApprovalRequest

A transaction approval request is used to log a request for a TransactionSignature under the specified key. The key must be owned by the UserId, and the transaction must involve the specified AssetId.

TransactionSignature

Signature on a Transaction under a digital asset key.