Struct lock_keeper::types::audit_event::AuditEvent

pub struct AuditEvent {
    pub audit_event_id: i64,
    pub account_id: AccountId,
    pub request_id: Uuid,
    pub key_id: Option<KeyId>,
    pub timestamp: OffsetDateTime,
    pub client_action: ClientAction,
    pub status: EventStatus,
}

A single entry that specifies the actor, action, outcome, and any related key for a logged audit event. We expect database implementors to create AuditEvent instances for us. So we make all fields public.

Fields
audit_event_id: i64account_id: AccountIdrequest_id: Uuidkey_id: Option<KeyId>timestamp: OffsetDateTime

We use [OffsetDateTime] as this is compatible with SQLx. Easily convertible to a postgres’ TIMESTAMPTZ type.

client_action: ClientActionstatus: EventStatus
Implementations

impl AuditEvent

pub fn request_id(&self) -> &Uuid

pub fn action(&self) -> ClientAction

pub fn key_id(&self) -> Option<&KeyId>

pub fn date(&self) -> OffsetDateTime

pub fn status(&self) -> EventStatus