Module lock_keeper::types::operations

Types related to server operations and the protocols they execute.

Modules

authenticate

create_storage_key

generate

get_user_id

import

logout

register

remote_generate

remote_sign_bytes

retrieve_audit_events

retrieve_secret

retrieve_server_encrypted_blob

retrieve_storage_key

store_server_encrypted_blob

Structs

ClientActionIter

An iterator over the variants of Self

RequestMetadata

Metadata attached to each request to the server. Note that the request ID is an ID for an entire operation, not each ClientAction that the operation is composed of.

Enums

ClientAction

Options for actions the Lock Keeper client can take. We explicitly assign discriminant values to this type, as it will be stored in the our database.

Traits

ConvertMessage

Converts a serializable Rust type to and from the RPC Message type.