Crate lock_keeper

Cryptography, protocols, and other shared types and context used by multiple entities in the Lock Keeper digital asset management system.

⚠️ Usage: This crate should not be used directly by developers outside of Bolt Labs. It is designed as a dependency for other crates in the ecosystem, including lock-keeper-key-server and lock-keeper-client.

Re-exports

pub use error::LockKeeperError;

Modules

config

Configuration types

constants

Constants that are shared between other crates in this workspace. Crate-specific constants should go in their respective crates.

crypto

Application-specific cryptographic types and operations.

error

Error type for all errors returned to code outside of this crate.

infrastructure

Modules that facilitate basic client-server operations

rpc

types

Type definitions that are shared between crates but have little to no logic.

Macros

impl_authenticated_message_conversion

Generates TryFrom implementations to and from the Message type for a given list of types that should only be sent authenticated.

impl_message_conversion

Generates TryFrom implementations to and from the Message type for a given list of types.

impl_message_conversion_internal

Common implementation for impl_message_conversion variants.