Module lock_keeper::crypto

Application-specific cryptographic types and operations.

Defines and implements keys and secret types, and the appropriate transformations between them. Public functions here are mostly wrappers around multiple low-level cryptographic steps.

Modules

arbitrary_secret 🔒

generic 🔒

signing_key 🔒

Structs

Encrypted

A ciphertext representing an object of type T, encrypted under the ChaCha20Poly1305 scheme for authenticated encryption with associated data (AEAD).

KeyId

Universally unique identifier for a stored secret or signing key.

MasterKey 🔒

The master key is a default-length symmetric encryption key for an AEAD scheme.

OpaqueExportKey

An export key is secure key material produced as client output from OPAQUE.

OpaqueSessionKey

A session key is produced as shared output for client and server from OPAQUE.

Secret

An arbitrary secret.

SigningKeyPair

An ECDSA signing key pair, including a public component for verifying signatures, a private component for creating them, and context about the key pair.

StorageKey

A storage key is a default-length symmetric encryption key for an AEAD scheme. The storage key is used to encrypt stored secrets and signing keys.

Enums

CryptoError

Errors that arise in the cryptography module.